代码提交1128-1

+++ dubboService/src/main/java/com/github/provider/dao/UserDiaryMapper.java
@@ -12,7 +12,7 @@ public interface UserDiaryMapper {
 
     int insertSelective(UserDiary record);
 
-    UserDiary selectByPrimaryKey(Integer id);
+    UserDiary selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(UserDiary record);
